requirements.txt
matplotlib==3.3.3
numpy==1.19.2
pandas==1.1.2
python-dotenv==0.15.0
requests==2.25.1
XlsxWriter==1.3.7
